集成 bosonnlp 关键词提取Api(4)

+++ database/migrations/2017_01_08_140733_create_article_key_words_table.php
@@ -17,7 +17,7 @@ class CreateArticleKeyWordsTable extends Migration
             $table->increments('id');
             $table->integer('key_words_id')->default(0)->unsigned();
             $table->integer('article_id')->default(0)->unsigned();
-            $table->integer('weight')->default(0)->unsigned();
+            $table->double('weight', 17, 16)->default(0)->unsigned();
             $table->timestamps();
 
             $table->foreign('article_id')->references('id')->on('articles')
